Agentic-ai-2-main/Agentic-ai-2-main/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
pysqlite3==0.5.0  # Optional: for better SQLite support

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# LangGraph (Agentic AI Orchestration)
# Note: If LangGraph is not installed, the system will automatically use a simple orchestration fallback
langgraph>=0.0.20
langchain>=0.1.0
langchain-core>=0.1.10

# Optional: For LLM integration (if you want to use actual LLM for insights)
# openai==1.3.0
# langchain-openai==0.0.2

# CORS (handled by FastAPI)

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

